Fixed script.

+++ start_container.sh
@@ -31,18 +31,20 @@ scratch_max_size=$((512*1024*1024*1024)) # 512GB
 echo Info: $(date)
 echo Info: image=$repo:$tag, hostname=$hostname, ip=$ip, user=$user, name=$name
 
-# Check current container
 container_exist=0
-for id in $(docker ps -a --filter "name=user-" --format "{{.ID}}") ; do
-  tmp=$(docker ps -a --filter "id=$id" --format "{{.Names}}")
-  if [[ $tmp != "$name" ]] ; then
-    echo Info: Stop user container: id=$id, name=$tmp
-    docker rm -f $id > /dev/null
-  else
-    echo Info: User container is already running: id=$id, name=$tmp
-    container_exist=1
-  fi
-done
+if [[ $name != "maintenance" ]] ; then
+  # Check current container and stop
+  for id in $(docker ps -a --filter "name=user-" --format "{{.ID}}") ; do
+    tmp=$(docker ps -a --filter "id=$id" --format "{{.Names}}")
+    if [[ $tmp != "$name" ]] ; then
+      echo Info: Stop user container: id=$id, name=$tmp
+      docker rm -f $id > /dev/null
+    else
+      echo Info: User container is already running: id=$id, name=$tmp
+      container_exist=1
+    fi
+  done
+fi
 
 # Start container
 if [ $container_exist -eq 0 ] ; then
